UserPerm: fix handling of alloced_gids in deep_copy
When we allocate a new UserPerm object with an existing one as a
template, we need to set alloced_gids to false so that deep_copy
won't try to delete the garbage pointer in the gids slot.

Also, deep_copy is a little ambiguous (to my eye) as it references
"this" on some fields and "a" on the others. Make it universally use
"a" since we're passing that in as a parameter.

Signed-off-by: Jeff Layton <jlayton@redhat.com>