mds: bump truncate seq when fscrypt_file changes
The MDS and client use a sequence number to logically order events that
truncate a file to a smaller size. Truncates to larger (or equal) size
don't require it however so the MDS doesn't increment the sequence in
that case.

This is problematic for the client storing the size in fscrypt_file. If
the truncate results in the last crypto block not changing size then the
classic size field and sequence won't change while fscrypt_file does.

To ensure that the client picks up the new size, increment the sequence
number when fscrypt_file changes and the MDS-apparent size does not.

Signed-off-by: Jeff Layton <jlayton@redhat.com>